extend functionality to read from configuration file
This is a first version only, to be improved. Prior to setting options with the command line, some
options can be read from the file '.agrc' located in user's home directory.

The content of the .agrc file consists of the ini-file syntax and may look like

```
[color]
linenumber = green
match = bold red
filename = magenta

[search]
case = sensitive

[output]
pager = less -FRSX
```
With this config file, ag performs a case-sensitive search where matched patterns are printed in
bold red, belonging line numbers in green and paths in magenta color. The output is processed with
'less -FRSX' as a pager.